# api-docs.yaml
openapi: 3.1.0
info:
  title: OpenAPI definition
  version: v0
servers:
- url: https://localhost:8443
  description: Generated server url
paths:
...
  /api/nutritions/{id}:
    get:
      tags:
      - nutrition-rest-controller
      summary: Get a nutrition by ID
      operationId: getNutrition
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: Nutrition found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NutritionDTO"
        "400":
          description: Bad request - Invalid parameters
        "401":
          description: Unauthorized access - Authentication is required
        "403":
          description: Forbidden - You don't have permission to access
        "404":
          description: Nutrition not found
        "500":
          description: Internal server error
    put:
      tags:
      - nutrition-rest-controller
      summary: Edit a nutrition by ID
      operationId: updateNutrition
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NutritionDTO"
        required: true
      responses:
        "200":
          description: Nutrition updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NutritionDTO"
        "400":
          description: Bad request - Invalid input data
        "401":
          description: Unauthorized access
        "403":
          description: Forbidden - You don't have permission to edit
        "404":
          description: Nutrition not found
        "500":
          description: Internal server error
    delete:
      tags:
      - nutrition-rest-controller
      summary: Delete a nutrition by ID
      operationId: deleteNutrition
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: Nutrition deleted successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NutritionDTO"
        "401":
          description: Unauthorized access
        "403":
          description: Forbidden - You don't have permission to delete
        "404":
          description: Nutrition not found
        "500":
          description: Internal server error
  /api/foods/{foodId}:
    get:
      tags:
      - food-rest-controller
      summary: Get a food by ID
      operationId: getTrainingByIdd
      parameters:
      - name: foodId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: Food found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FoodDTO"
        "400":
          description: Bad request - Invalid parameters
        "401":
          description: Unauthorized access - Authentication is required
        "403":
          description: Forbidden - You don't have permission to access
        "404":
          description: Food not found
        "500":
          description: Internal server error
    put:
      tags:
      - food-rest-controller
      summary: Replace a food
      operationId: replacePost
      parameters:
      - name: foodId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FoodDTO"
        required: true
      responses:
        "200":
          description: Food replaced
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FoodDTO"
        "400":
          description: Bad request - Invalid parameters
        "401":
          description: Unauthorized access - Authentication is required
        "403":
          description: Forbidden - You don't have permission to access
        "404":
          description: Food not found
        "500":
          description: Internal server error
    delete:
      tags:
      - food-rest-controller
      summary: Delete a food
      operationId: deleteFood
      parameters:
      - name: foodId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: Food deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FoodDTO"
        "400":
          description: Bad request - Invalid parameters
        "401":
          description: Unauthorized access - Authentication is required
        "403":
          description: Forbidden - You don't have permission to access
        "404":
          description: Food not found
        "500":
          description: Internal server error
...
  /api/foods/{id}:
    delete:
      tags:
      - food-rest-controller
      operationId: deleteFood_1
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
components:
  schemas:
...
    NutritionDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        type:
          type: string
        userId:
          type: integer
          format: int64
        listIdsFood:
          type: array
          items:
            type: integer
            format: int64
    FoodDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        calories:
          type: integer
          format: int32
        type:
          type: string
        listIdsNutrition:
          type: array
          items:
            type: integer
            format: int64
...